feat:调整output的金额相关类型

See merge request ckb-explorer/ckb-explorer-web!157

+++ src/main/java/com/ckb/explorer/domain/dto/PendingCellInputDto.java
@@ -1,5 +1,6 @@
 package com.ckb.explorer.domain.dto;
 
+import java.math.BigInteger;
 import lombok.Data;
 
 @Data
@@ -9,12 +10,12 @@ public class PendingCellInputDto {
   /**
    * 单元格容量
    */
-  private Long capacity;
+  private BigInteger capacity;
 
   /**
    * 占用的容量
    */
-  private Long occupiedCapacity;
+  private BigInteger occupiedCapacity;
 
   private byte[] lockCodeHash;
